Checkpoint 1: Literature review
The literature review begins by introducing the reader to the real-time strategy (RTS) subgenre. It then continues by expanding on artificial intelligence (AI) and its use in video games in general, mentioning its first popular appearance in Pac-Man and bringing up the importance of AI in games. Following that, it categorises AI techniques into two; deterministic and non-deterministic, gives a brief description of both, and indicates how RTS games generally fall under the non-deterministic category due to their stochastic nature. Subsequently, it gives an overview of the current work on RTS game AI by discussing what has been achieved so far. It then lists several machine learning methods and algorithms, going into detail about reinforcement learning and Proximal Policy Optimization (PPO) specifically, which was the algorithm of choice for this study. Finally, it refers to several works that relate to this study and outlines what was achieved within them.
For the literature review, it was made sure that past tense and passive writing are used throughout. Statements by other researchers and other works were supported by in-text citations and references. It is worth adding that some explanations were also backed up by diagrams for clarification. The word count is approx. 4200, which exceeds the recommended length. This section requires some fine-tuning for it to be up to standard.