Checkpoint 2: Methodology
The methodology introduction starts by outlining the study’s focus. It then continues by describing how the study attempted to fill the gaps in current research. It also states that two prototypes were developed, along with the approaches that were taken to provide results. Following that, the methodology specifies and describes the technologies and tools that were used to conduct the study, providing justifications for their use. Next, it expands on the prototypes by describing the steps that were taken to develop them while also outlining the components they comprise. This section contains several figures that back up and clarify its contents. It then continues by describing the prototype’s design limitations and concludes by giving detailed descriptions of the chosen data collection methods, mentioning the use of graphs, test cases and qualitative interviews.
Like in the literature review, it was made sure that past tense and passive writing are used throughout. The word count for this section is approx. 3300, which exceeds the recommended length. This section requires some fine-tuning for it to be up to standard.