Checkpoint 4: Results and discussion
This section starts by mentioning that it will be describing how the study’s results will correlate with the research questions and hypothesis.
Following that, it presents the quantitative data collected from both approaches in the form of graphs; each graph will be evaluated in detail. Moreover, it also compares the two approaches by analysing the differences in statistics. These comparisons are backed by several figures. Most importantly, it then states that the data cannot be compared to that from other studies due to the difference in approaches while explaining why.
This section then continues by outlining and discussing the tests that were done for this study by presenting them in the form of test cases in tables. Whenever a result was not as expected, the potential reasons were given along with their evaluations. Some of the test cases are discussed in further detail.
Finally, this section outlines the results collected from the three 1-on-1 interviews that were conducted for this study. It lists the observations that were made by the participants, categorising them by approach. The interview transcripts were included in an Appendix to which there are several references that back these observations. Moreover, the observations are also backed by screenshots taken directly from the videos that were recorded for the interviews. It concludes with an analysis of the contrast within said observations.
Like in the other sections, it was made sure that past tense and passive writing are used throughout. The word count for this section is approx. 3200. This section requires some fine-tuning for it to be up to standard.