Checkpoint 5: Conclusion, introduction, abstract (still need to be done)
The introduction will contain the following:
-	The definition of the problem: There is no research that addresses the use of reinforcement learning (via Unity and ML-Agents) for pathfinding in RTS games.
-	Motivation behind the research: The interest in video game development and artificial intelligence, specifically with the use of the Unity game engine.
-	A definition of the research objectives: Making use of Unity and ML-Agents to train agents that can adapt to a dynamic environment.
-	A section describing the chapters in the study.

The conclusion will contain the following:
-	Summary of the findings from both approaches.
-	Recommendations for future research.
-	A critical evaluation of both approaches with emphasis on their strengths and weaknesses.

The abstract will consist of a summary with highlights from the study. It will start by giving a brief introduction of AI and its use in games. It will then continue by stating the study’s research objectives along with what was done to achieve them.

These still need to be done; it will be made sure that a passive style of writing is used. References will be provided in context.